+++ doc/CHANGELOG.md
@@ -41,10 +41,10 @@
 - [x] `NystromSolver` improvements:
   - [x] BREAKING CHANGE: handle logarithmic functions as instances of `DirichletTrace` class
   - [x] move `jacobi_preconditioner()` static method to a separate module
-- [ ] `locfun` module improvements:
-  - [ ] use `DirichletTrace` objects for traces in `antilap` module
+- [x] `locfun` module improvements:
+  - [x] use `DirichletTrace` objects for traces in `antilap` module
   - [x] move contents of `antilap` and `d2n` subpackages to `locfun` root
-  - [ ] remove `log_terms` and `log_antilap` modules
+  - [x] remove `log_terms` and `log_antilap` modules
 - [ ] refactor `GlobalFunctionSpace`:
   - [ ] build from a collection of `GlobalFunction` objects, just like `LocalFunctionSpace` is built from a collection of `LocalFunction` objects
   - [ ] precompute `LocalFunctionSpace` objects